ref(init): make room for 2nd version
this is the start of the 2nd version of achievibit. I've switched to nest.js and angular (even though front-end might change)

BREAKING CHANGE: From this point on, the application is going to work differently, so this is definitely a breaking change that should start the 2nd version. Even after feature parity, the structure will be different in this new version (api route moves, client side will be fancier, etc.).